state capitalism
its a mixed bag, I'm not sure if its too good or too bad. if you have a highly productive industry it'll bring you tons of money but it costs loads of bureaucracy. I think it's good for nations with a labor surplus and/or nations that are not very industrialized (which is the goal).